+++ android/src/main/java/com/picker/Utils.java
@@ -1,12 +1,18 @@
-package com.picker;
+package com.chavesgu.images_picker;
 
 import android.content.pm.ActivityInfo;
 
 import com.luck.picture.lib.PictureSelectionModel;
+import com.luck.picture.lib.PictureSelector;
 import com.luck.picture.lib.config.PictureConfig;
+import com.luck.picture.lib.config.PictureMimeType;
 import com.luck.picture.lib.language.LanguageConfig;
+import com.yalantis.ucrop.view.OverlayView;
 
+import java.lang.reflect.Array;
 import java.util.HashMap;
+import java.util.List;
+import java.util.UUID;
 
 public class Utils {
     public static PictureSelectionModel setPhotoSelectOpt(PictureSelectionModel model, int count, double quality) {
@@ -24,16 +30,17 @@ public class Utils {
                 .isGif(true)
                 .isEnableCrop(false)
                 .isCompress(false)
-                .compressFocusAlpha(true)
                 .minimumCompressSize(100)
                 .isReturnEmpty(false)
                 .isAndroidQTransform(true)
                 .setRequestedOrientation(ActivityInfo.SCREEN_ORIENTATION_PORTRAIT)
                 .isOriginalImageControl(false)
                 .isMaxSelectEnabledMask(true)
-//                .cameraFileName("images_picker_camera")
-//                .renameCompressFile("images_picker_compress")
-//                .renameCropFileName("images_picker_crop")
+                .setCameraImageFormat(PictureMimeType.JPEG)
+                .setCameraVideoFormat(PictureMimeType.MP4)
+                .renameCompressFile("image_picker_compress_"+UUID.randomUUID().toString()+".jpg")
+                .renameCropFileName("image_picker_crop_"+UUID.randomUUID().toString()+".jpg")
+//                .cameraFileName("image_picker_camera_"+UUID.randomUUID().toString()+".jpg")
         ;
         if (quality > 0) {
             model.isCompress(true).compressQuality((int) ((double) quality * 100));
@@ -44,7 +51,7 @@ public class Utils {
     public static PictureSelectionModel setCropOpt(PictureSelectionModel model, HashMap<String, Object> opt) {
         model
                 .isEnableCrop(true)
-                .freeStyleCropEnabled(true)
+                .freeStyleCropMode(OverlayView.FREESTYLE_CROP_MODE_ENABLE)
                 .circleDimmedLayer(opt.get("cropType").equals("CropType.circle"))
                 .showCropFrame(!opt.get("cropType").equals("CropType.circle"))
                 .showCropGrid(false)
@@ -53,7 +60,6 @@ public class Utils {
                 .isDragFrame(true)
                 .hideBottomControls(false)
                 .isMultipleSkipCrop(true)
-                .compressFocusAlpha(true)
                 .cutOutQuality((int) ((double) opt.get("quality") * 100));
         if (opt.get("aspectRatioX") != null) {
             model.isDragFrame(false);